--- a/benchmarks/linear_algebra/blas/level3/sgemm/gpu/CMakeLists.txt
+++ b/benchmarks/linear_algebra/blas/level3/sgemm/gpu/CMakeLists.txt
@@ -13,6 +13,6 @@ set_target_properties(${wrapper_name} PROPERTIES COMPILE_FLAGS -O3)
 target_link_libraries(${wrapper_name} tiramisu ${HalideLib} ${ISLLib} ${LINK_FLAGS} cuda_wrapper ${CUDA_LIBRARIES} ${CUDA_CUBLAS_LIBRARIES})
 
 add_custom_target(run_${benchmark_name} COMMAND ${wrapper_name} 100 0 DEPENDS ${wrapper_name})
-add_custom_target(run_${benchmark_name}_nvprof COMMAND LD_LIBRARY_PATH=${CUDA_TOOLKIT_ROOT_DIR}/lib64 ${CUDA_TOOLKIT_ROOT_DIR}/bin/nvprof --print-gpu-trace $<TARGET_FILE:${wrapper_name}> 1 0 DEPENDS ${wrapper_name})
-add_custom_target(run_${benchmark_name}_nvprof2 COMMAND LD_LIBRARY_PATH=${CUDA_TOOLKIT_ROOT_DIR}/lib64 ${CUDA_TOOLKIT_ROOT_DIR}/bin/nvprof $<TARGET_FILE:${wrapper_name}> 100 0 DEPENDS ${wrapper_name})
+add_custom_target(run_${benchmark_name}_nvprof COMMAND LD_LIBRARY_PATH=${CUDA_TOOLKIT_ROOT_DIR}/lib64 ${CUDA_TOOLKIT_ROOT_DIR}/bin/nvprof --profile-from-start off --print-gpu-trace $<TARGET_FILE:${wrapper_name}> 1 0 DEPENDS ${wrapper_name})
+add_custom_target(run_${benchmark_name}_nvprof2 COMMAND LD_LIBRARY_PATH=${CUDA_TOOLKIT_ROOT_DIR}/lib64 ${CUDA_TOOLKIT_ROOT_DIR}/bin/nvprof --profile-from-start off $<TARGET_FILE:${wrapper_name}> 100 0 DEPENDS ${wrapper_name})
 add_custom_target(run_${benchmark_name}_correctness COMMAND ${wrapper_name} 0 1 DEPENDS ${wrapper_name})