Adding prefetching and global memory coalescing on warp level to GPU Gemm.

I realized the kernel run time doesn't reach steady state for 10-15 runs. So I've increased the warm up runs. I've also fixed a global memory coalescing issue I found in profiler. Here are the latest numbers (GPU time, 3072x3072):

|                 | Pascal 4 | Kepler 80 (Salike) |
|-----------------|----------|--------------------|
| Old             | 20ms     | 73ms               |
| After Prefetch  | 19.2ms   | 74.5ms             |
| **After Coalescing** | **18.3ms**   | **62.5ms**             |
| cuBLAS          | 10.8ms   | 28ms               |

Note that unlike P4, prefetching decreases the performance slightly on K80. But combined, we get improvements of 9% on P4 and 15% on K80 with this diff.

**Next steps:** Now profiler shows some issues with shared memory. Also disassembling cuBLAS shows that they are using bulk memory operations and more unrolling.